[Refactor] Remove old mctest, replace with updated version
