+++ tools/Python/mctest/mctest.py
@@ -1,5 +1,9 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
+from dataclasses import dataclass
+from typing import Union
+from pathlib import Path
+
 import logging
 import argparse
 import json
@@ -20,288 +24,310 @@ from mccodelib import utils, mccode_config
 #
 # Functionality
 #
+def config_override(version: str):
+    flavour = mccode_config.configuration['MCCODE']
+    return Path(__file__).parent.joinpath("mccodelib", f"{flavour}-test", version).resolve()
 
-def create_instr_test_objs(sourcefile, localfile, header):
-    ''' returns a list containing one initialized test object pr %Example within the instr file '''
-    tests = []
-    ms = re.findall("\%Example:([^\n]*)Detector\:([^\n]*)_I=([0-9.+-e]+)", header)
-    if len(ms) > 0:
-        testnb = 1
-        for m in ms:
-            parvals = m[0].strip()
-            detector = m[1].strip()
-            targetval = float(m[2].strip())
-            tests.append(InstrExampleTest(sourcefile, localfile, parvals, detector, targetval, testnb))
-            testnb = testnb + 1
-    else:
-        tests.append(InstrExampleTest(sourcefile, localfile))
-    return tests
 
-class InstrExampleTest:
-    ''' instruent test house keeping object '''
-    def __init__(self, sourcefile, localfile, parvals=None, detector=None, targetval=None, testnb=0):
+def regular_expression(key: str):
+    from re import compile
+    parameters = "(?P<parameters>[^\n]*)"
+    monitor = "(?P<monitor>[^\n]+)_I"
+    not_end_comment_or_end_line = r"(?:\*(?!/)|[^\*\n])"
+    inner = f"(?:{not_end_comment_or_end_line}+|\\[{not_end_comment_or_end_line}+\\])"
+    intensity = f"(?P<intensity>{inner})"
+    error = f"\\s*\\2_ERR\\s*=\\s*(?P<error>{inner})"
+    case = f"{key}:{parameters}Detector:\\s*{monitor}\\s*=\\s*{intensity}(?:{error})?"
+    case = f"(?P<case>{case})"
+    return compile(case)
+
+
+def filter_comments(contents: str):
+    """Concatenate across continuation lines, then return only the comment line(s)"""
+    from re import compile, DOTALL, finditer
+    # Concatenate continuation lines
+    contents = ' '.join(contents.split('\\\n'))
+    # Match all C-style comments: blocks and lines
+    comment = compile(r"(?:/\*(?:\*(?!/)|[^\*])*\*/|//[^\n]*\n)", DOTALL)
+    # recombine string with only the comments
+    comments = '\n'.join([x.group(0) for x in finditer(comment, contents)])
+    return comments
+
+TEST_CASE_REGEX = regular_expression("%Example")
+
+@dataclass
+class InstrExampleTestCase:
+    case: str
+    parameters: str
+    monitor: str
+    intensity:  Union[float, list[float]]
+    error:  Union[None, float, list[float]] = None
+    result: Union[None, float, list[float]] = None
+    ran: Union[None, bool] = None
+    runtime: Union[None, float] = None
+    errmsg: Union[None, str] = None
+
+    def __post_init__(self):
+        self.parameters = self.parameters.strip()
+        if isinstance(self.intensity, str):
+            self.intensity = eval(self.intensity, None, None)
+        if self.error is not None and isinstance(self.error, str):
+            self.error = eval(self.error, None, None)
+        if self.error is not None and not isinstance(self.error, type(self.intensity)):
+            raise RuntimeError("Intensity and errors must evaluate to the same type!")
+        # Stash-away whether this test is a scan
+        self._is_scan = '-N' in self.parameters and (hasattr(self.intensity, '__len__') and len(self.intensity) > 0)
+
+    @classmethod
+    def scan(cls, header):
+        from re import finditer
+        tests = [cls(**match.groupdict()) for match in finditer(TEST_CASE_REGEX, header)]
+        return tests
+
+    def json_repr(self):
+        return dict(parvals=self.parameters, detector=self.monitor, targetval=self.intensity, testval=self.result,
+                    didrun=self.ran, runtime=self.runtime, errmsg=self.errmsg)
+
+    def run(self, working_directory, cmd: str, no_particles: Union[int,float], test_no: int):
+        logging.debug(f"Run test case {self.case}\nwith parameters {self.parameters} and monitor {self.monitor}")
+        # cmd should have everything up through " -s 1000 [filename]"
+        global openacc, mpi
+        import time
+        from pathlib import Path
+        t0 = time.time_ns()
+        cmd += f" {self.parameters} -n{no_particles}"
+        if mpi is not None:
+            if openacc is True:
+                cmd += " --openacc"
+            cmd += f" --mpi={mpi}"
+        run_log = working_directory.joinpath(f"run_stdout_{test_no}.txt").resolve()
+        cmd += f" -d {test_no} &> {run_log}"
+        code = utils.run_subtool_noread(cmd, cwd=working_directory)
+        t1 = time.time_ns()
+        output_path = Path(working_directory).joinpath(f"{test_no}").resolve()
+        self.result = self.extract_values(output_path)[0]  # TODO Use the error and number of particles information?
+        self.ran = (code != 0 or output_path.joinpath("mccode.sim").is_file()) and self.result is not None
+        if not self.ran:
+            with open(run_log, 'r') as logfile:
+                self.errmsg = logfile.read()
+        self.runtime = (t1 - t0) / 1e9  # nanoseconds to seconds
+        return output_path
+
+    def extract_values(self, directory):
+        # Select begin XXX ... end XXX blocks:
+        sim_ex = re.compile(r"begin (?P<environment>[a-zA-Z]+):?(?P<contents>(?:(?!begin)|.)+)end \1", re.DOTALL)
+
+        def get_monitor_filename(root):
+            with open(root.joinpath('mccode.sim')) as file:
+                string = file.read()
+            for match in re.finditer(sim_ex, string):
+                if match.group('environment') == 'data':
+                    lines = match.group('contents').split('\n')
+                    data = {k.strip(): v.strip() for k, v in [n.split(': ') for n in lines if ': ' in n]}
+                    if 'component' in data and data['component'] == self.monitor:
+                        monitor_path = root.joinpath(data.get('filename', self.monitor + '.dat'))
+                        return monitor_path if monitor_path.is_file() else None
+            return None
+
+        def get_monitor_values(root: Path):
+            file_path = get_monitor_filename(root)
+            if file_path is None:
+                logging.debug(f"{directory} {self.monitor} file is 'None' -- did the simulation actually run?")
+                return None, None, None
+            if not file_path.is_file():
+                logging.debug(f"Monitor file {file_path} does not exist")
+                return None, None, None
+            with open(file_path, 'r') as monitor_file:
+                monitor_lines = monitor_file.readlines()
+            monitor_lines = list(filter(lambda x: '# values:' in x, monitor_lines))
+            if len(monitor_lines) != 1:
+                logging.debug(f"{len(monitor_lines)} values in {file_path}")
+            if len(monitor_lines) == 0:
+                return None, None, None
+            return [float(x) for x in monitor_lines[0].strip().split(': ')[1].split()]
+
+        points = 0
+        if self._is_scan:
+            with open(directory.joinpath('mccode.sim')) as file:
+                lines = file.readlines()
+            lines = list(filter(lambda x: "Numpoints:" in x, lines))
+            points = int(lines[0].strip().split(': ')[1]) if len(lines) == 1 else 0
+
+        if points:
+            values, errors, counts = zip(*[get_monitor_values(directory.joinpath(f'{p}')) for p in range(points)])
+            return values, errors, counts
+
+        return tuple(get_monitor_values(directory))
+
+
+class InstrExampleTests:
+    def __init__(self, sourcefile, localfile, max_filename=160):
         self.sourcefile = sourcefile
         self.localfile = localfile
-        self.instrname = splitext(basename(sourcefile))[0]
-        self.testnb = testnb
-        
-        self.parvals = parvals
-        self.detector = detector
-        self.targetval = targetval
-        self.testval = None
-
+        with open(localfile, 'r') as file:
+            contents = file.read()
+        self.tests = InstrExampleTestCase.scan(filter_comments(contents))
         self.compiled = None
-        self.compiletime = None
-        self.didrun = None
-        self.runtime = None
-        self.errmsg = None
-    def get_json_repr(self):
-        return {
-            "displayname"  : self.get_display_name(),
-            "sourcefile"   : self.sourcefile,
-            "localfile"    : self.localfile,
-            "instrname"    : self.instrname,
-            "testnb"       : self.testnb,
+        self.compiletime = 0
 
-            "parvals"      : self.parvals,
-            "detector"     : self.detector,
-            "targetval"    : self.targetval,
-            "testval"      : self.testval,
+        self.log_name = "{fn:<{fm}}".format(fn=str(localfile), fm=f"{max_filename:d}s")[:max_filename]
+        no_test = "TEST" if len(self.tests) < 2 else f"TEST {len(self.tests)}"
+        if len(self.tests) < 1:
+            no_test = "NO TEST"
+        logging.debug(f"{self.log_name}: {no_test}")
 
-            "compiled"     : self.compiled,
-            "compiletime"  : self.compiletime,
-            "didrun"       : self.didrun,
-            "runtime"      : self.runtime,
-            "errmsg"       : self.errmsg,
-        }
-    def save(self, infolder):
-        text = json.dumps(self.get_json_repr())
-        f = open(join(infolder, self.get_display_name()) + ".json", 'w').write(text)
-    def get_display_name(self):
-        if self.testnb > 1:
-            return self.instrname + "_%d" % self.testnb
-        else:
-            return self.instrname
+    @property
+    def instrname(self):
+        return splitext(basename(self.sourcefile))[0]
 
-class LineLogger():
-    ''' log lines to memory, then save to disk '''
+    def json_repr(self):
+        def base(extend: dict, n=None):
+            source = str(self.sourcefile)
+            local = str(self.localfile)
+            d = dict(displayname=self.instrname if n is None else f"{self.instrname}_{n}",
+                     sourcefile=source, localfile=local, instrname=self.instrname,
+                     testnb=1 if n is None else n, compiled=self.compiled, compiletime=self.compiletime)
+            d.update(extend)
+            return d
+        return [base(x.json_repr(), n if len(self.tests) > 1 else None) for n, x in enumerate(self.tests)]
+
+    def json_dict(self):
+        return {x['displayname']: x for x in self.json_repr()}
+
+    def save(self, directory):
+        for content in self.json_repr():
+            with open(directory.joinpath(content['displayname'] + '.json'), 'w') as file:
+                file.write(json.dumps(content))
+
+    def compile(self, version=None):
+        binary = self.localfile.with_suffix(f".{mccode_config.platform['EXESUFFIX']}")
+        if binary.is_file():
+            self.compiled = True
+        elif self.compiled is None:
+            logging.info(f"{binary} does not exist (yet)")
+            from pathlib import Path
+            import time
+            global mpi, openacc
+            t0 = time.time_ns()
+            cmd = mccode_config.configuration["MCRUN"]
+            if version is not None:
+                cmd += f" --override-config={config_override(version)}"
+            if openacc:
+                cmd += " --openacc "
+            if mpi:
+                cmd += " --mpi=1 "
+            cmd += f" --verbose -c -n0 {self.localfile.resolve()} &> compile_stdout.txt"
+            utils.run_subtool_noread(cmd, cwd=self.localfile.parent)
+            t1 = time.time_ns()
+            self.compiled = binary.is_file()
+            self.compiletime = (t1 - t0) / 1e9  # convert from nanoseconds to seconds
+            if self.compiled:
+                logging.info(f"{self.log_name}: {self.compiletime:3.2f} s")
+            else:
+                logging.info(f"{self.log_name}: COMPILE ERROR using\n{cmd}")
+
+    def test(self, version, no_particles):
+        if not self.compiled:
+            logging.info(f"{self.log_name} NO COMPILE")
+            return
+        if len(self.tests) < 1:
+            logging.info(f"{self.log_name} NO TEST")
+            return
+
+        cmd = mccode_config.configuration["MCRUN"]
+        if version:
+            cmd += f" --override-config={config_override(version)}"
+        cmd += f" -s 1000 {self.localfile.resolve()}"
+
+        for number, to_test in enumerate(self.tests):
+            to_test.run(self.localfile.parent, cmd, no_particles, number)
+            msg = f"{to_test.runtime:3.2f} s" if to_test.ran else "RUNTIME ERROR"
+            logging.info(f"{self.log_name}({number+1:2d}/{len(self.tests):2d}): {msg}")
+
+        self.save(self.localfile.parent)
+
+
+class LineLogger:
+    """ log lines to memory, then save to disk """
     def __init__(self):
         self.lst = []
-    def logline(self, addstr):
-        self.lst.append(addstr)
+
+    def logline(self, line):
+        self.lst.append(line)
+
     def save(self, filename):
-        text = "\n".join(self.lst) + "\n"
-        f = open(filename, 'w').write(text)
-    def find(self, searchstr):
-        for l in self.lst:
-            if re.search(searchstr, l):
+        with open(filename, 'w') as file:
+            file.write('\n'.join(self.lst) + '\n')
+
+    def find(self, sub):
+        for line in self.lst:
+            if re.search(sub, line):
                 return True
         return False
 
-def _monitorname_filename_match(dfolder, monname):
-    '''
-    mccode.sim is organized in sections, e.g. data sections are coded by from "begin data" to "end data",
-    within which "  component:" and "  filename:" tags may be available.
-    
-    returns the filename or None
-    '''
-    look_for_filename = False
-    lns = open(join(dfolder, "mccode.sim")).read().splitlines()
-    for l in lns:
-        if re.match("  component: %s$" % monname, l):
-            # flag this data section 
-            look_for_filename = True
-        if look_for_filename:
-            m = re.match("\s*filename:\s+(.+)", l)
-            if m:
-                filename = m.group(1) 
-                if os.path.isfile(join(dfolder,filename)):
-                    return filename 
-            if re.match("end data", l):
-                # the filename can for 0D monitors be monname.dat
-                zeroDfilename = join(dfolder, monname + ".dat")
-                if os.path.isfile(zeroDfilename):
-                    return zeroDfilename
-                return None
 
-def extract_testvals(datafolder, monitorname):
-    '''
-    Extract monitor I (as well as Ierr and N) from results dir given monitor name.
-    
-    Returns an error string or a tuple containing (I, I_err, N), or None if datafile does not contain a "values" line.
-    '''
-    # get any filename in mccode.sim matching monitorname AKA test.detector 
-    filename = _monitorname_filename_match(datafolder, monitorname)
-    if filename is None:
-        return "ERROR: targetval could not be extracted from monitor %s" % (monitorname)
-
-    # extract tested target value from the monitor file
-    with open(join(datafolder, filename)) as fp:
-        while True:
-            l = fp.readline()
-            if not l:
-                break
-            m = re.match("# values: ([0-9+-e.]+) ([0-9+-e.]+) ([0-9]+)", l)
-            if m :
-                I = float(m.group(1))
-                I_err = float(m.group(2))
-                N = float(m.group(3))
-                return (I, I_err, N)
-                break
-
-def mccode_test(branchdir, testdir, limitinstrs=None, instrfilter=None, version=None):
-    ''' this main test function tests the given mccode branch/version '''
+def mccode_test(branchdir: Path, testdir: Path, limitinstrs=None, instrfilter=None, version=None):
+    """ this main test function tests the given mccode branch/version """
+    from pathlib import Path
+    from shutil import copy
 
     # copy instr files and record info
-    logging.info("Finding instruments in: %s" % branchdir)
-    instrs, _ = utils.get_instr_comp_files(join(branchdir, "examples"), recursive=True, instrfilter=instrfilter)
+    logging.info(f"Finding instruments in: {branchdir}\n")
+    instrs, _ = utils.get_instr_comp_files(branchdir.joinpath("examples"), recursive=True, instrfilter=instrfilter)
     instrs.sort()
 
     # limt runs if required
     if limitinstrs:
         instrs = instrs[:limitinstrs]
 
+    if len(instrs) == 0:
+        logging.info("No instruments found for testing! Check configuration")
+        exit(0)
+
     # max instr name length for pretty-output
-    maxnamelen = 0
-    for f in instrs:
-        l = len(basename(f)) - 5
-        if l > maxnamelen:
-            maxnamelen = l
+    maxnamelen = max([len(str(file)) for file in instrs])
 
     # create test objects and copy instrument files
     logging.info("Copying instruments to: %s" % testdir)
     tests = []
     for f in instrs:
-        # create the test foldr for this instrument
-        instrname = splitext(basename(f))[0]
-        instrdir = join(testdir, instrname)
-        mkdir(instrdir)
+        file = Path(f)
+        # create the test folder for this instrument
+        instrument_directory = Path(testdir).joinpath(file.stem)
+        # if instrument_directory.is_dir():
+        #     raise RuntimeError(f"The specified testing directory {instrument_directory} already exists!")
+        if not instrument_directory.is_dir():
+            instrument_directory.mkdir(parents=True)
 
         # create a new file with the instr text in it - e.g. a local copy of the instrument file
-        text = open(f, encoding='utf-8').read()
-        f_new = join(instrdir, basename(f))
-        open(f_new, 'w', encoding='utf-8').write(text)
+        new_file = instrument_directory.joinpath(file.name)
+        copy(file, new_file)
 
         # create a test object for every test defined in the instrument header
-        instrtests = create_instr_test_objs(sourcefile=f, localfile=f_new, header=text)
-        tests = tests + instrtests
-
-        # extract and record %Example info from text
-        numtests = len([t for t in instrtests if t.testnb > 0]) 
-        if numtests == 0:
-            formatstr = "%-" + "%ds: NO TEST" % maxnamelen
-            logging.debug(formatstr % instrname)
-        elif numtests == 1:
-            formatstr = "%-" + "%ds: TEST" % maxnamelen
-            logging.debug(formatstr % instrname)
-        else:
-            formatstr = "%-" + "%ds: TESTS (%d)" % (maxnamelen, numtests)
-            logging.debug(formatstr % instrname)
-
+        # (automatically logs test information at debug level)
+        instrument_tests = InstrExampleTests(file, new_file, maxnamelen)
+        tests.append(instrument_tests)
 
     # compile, record time
     global ncount, mpi, openacc
     logging.info("")
     logging.info("Compiling instruments [seconds]...")
     for test in tests:
-        # if binary exists, set compile time = 0 and continue
-        binfile = os.path.splitext(test.localfile)[0] + ".out"
-        if os.path.exists(binfile):
-            test.compiled = True
-            test.compiletime = 0
-        else:
-            if test.testnb > 0 or (not args.skipnontest):
-                log = LineLogger()
-                t1 = time.time()
-                cmd = mccode_config.configuration["MCRUN"]
-                if version:
-                    cmd = cmd + " --override-config=" + join(os.path.dirname(__file__), "..", "mccodelib", mccode_config.configuration["MCCODE"] + "-test",version)
-                if openacc:
-                    cmd = cmd + " --openacc "
-                if mpi:
-                    cmd = cmd + " --mpi=1 "
-                cmd = cmd + " --verbose -c -n0 %s &> compile_stdout.txt" % test.localfile
-                utils.run_subtool_noread(cmd, cwd=join(testdir, test.instrname))
-                t2 = time.time()
-                test.compiled = os.path.exists(binfile)
-                test.compiletime = t2 - t1
-
-                # log to terminal
-                if test.compiled:
-                    formatstr = "%-" + "%ds: " % maxnamelen + \
-                      "{:3d}.".format(math.floor(test.compiletime)) + str(test.compiletime-int(test.compiletime)).split('.')[1][:2]
-                    logging.info(formatstr % test.get_display_name())
-                else:
-                    formatstr = "%-" + "%ds: COMPILE ERROR using:\n" % maxnamelen
-                    logging.info(formatstr % test.instrname + cmd)
-            else:
-                logging.info("Skipping compile of " + test.instrname)
+        test.compile(version)
         # save (incomplete) test results to disk
-        test.save(infolder=join(testdir, test.instrname))
+        test.save(test.localfile.parent)
 
     # run, record time
     logging.info("")
     logging.info("Running tests...")
     for test in tests:
-        if not test.compiled:
-            formatstr = "%-" + "%ds:   NO COMPILE" % (maxnamelen+1)
-            logging.info(formatstr % test.instrname)
-            continue
-        
-        # runable tests have testnb > 0
-        if test.testnb <= 0:
-            formatstr = "%-" + "%ds:   NO TEST" % (maxnamelen+1)
-            logging.info(formatstr % test.get_display_name())
-            continue
-
-        # run the test, record time and runtime success/fail
-        t1 = time.time()
-        cmd = mccode_config.configuration["MCRUN"]
-        if mpi is not None:
-            if openacc is True:
-                if version:
-                    cmd = cmd + " --override-config=" + join(os.path.dirname(__file__), "..", "mccodelib", mccode_config.configuration["MCCODE"] + "-test",version)
-                cmd = cmd + " -s 1000 %s %s -n%s --openacc --mpi=%s -d%d &> run_stdout_%d.txt" % (test.localfile, test.parvals, ncount, mpi, test.testnb, test.testnb)
-            else:
-                if version:
-                    cmd = cmd + " --override-config=" + join(os.path.dirname(__file__), "..", "mccodelib", mccode_config.configuration["MCCODE"] + "-test",version)
-                cmd = cmd + " -s 1000 %s %s -n%s --mpi=%s -d%d &> run_stdout_%d.txt" % (test.localfile, test.parvals, ncount, mpi, test.testnb, test.testnb)
-        else:
-            if version:
-                    cmd = cmd + " --override-config=" + join(os.path.dirname(__file__), "..", "mccodelib", mccode_config.configuration["MCCODE"] + "-test",version)
-            cmd = cmd + " -s 1000 %s %s -n%s -d%d  &> run_stdout_%d.txt" % (test.localfile, test.parvals, ncount, test.testnb, test.testnb)
-        retcode = utils.run_subtool_noread(cmd, cwd=join(testdir, test.instrname))
-        t2 = time.time()
-        didwrite = os.path.exists(join(testdir, test.instrname, str(test.testnb), "mccode.sim"))
-        test.didrun = retcode != 0 or didwrite
-        test.runtime = t2 - t1
-
-        # log to terminal
-        if test.didrun:
-            formatstr = "%-" + "%ds: " % (maxnamelen+1) + \
-                "{:3d}.".format(math.floor(test.runtime)) + str(test.runtime-int(test.runtime)).split('.')[1][:2]
-            logging.info(formatstr % test.get_display_name())
-        else:
-            formatstr = "%-" + "%ds: RUNTIME ERROR" % (maxnamelen+1)
-            logging.info(formatstr % instrname + ", " + cmd)
-            continue
-
-        # test value extraction
-        extraction = extract_testvals(join(testdir, test.instrname, str(test.testnb)), test.detector)
-        if type(extraction) is tuple:
-            test.testval = extraction[0]
-        else:
-            test.testval = -1
-
-        # save test result to disk
-        test.testcomplete = True
-        test.save(infolder=join(testdir, test.instrname))
+        test.test(version, ncount)
 
     #    cpu type: cat /proc/cpuinfo |grep name |uniq | cut -f2- -d: 
     #    gpu type: nvidia-smi -L | head -1 |cut -f2- -d: |cut -f1 -d\(
 
     metalog = LineLogger()
-    utils.run_subtool_to_completion("cat /proc/cpuinfo |grep name |uniq | cut -f2- -d: | xargs echo", stdout_cb=metalog.logline)
+    utils.run_subtool_to_completion("cat /proc/cpuinfo |grep name | uniq | cut -f2- -d: | xargs echo", stdout_cb=metalog.logline)
     cpu_type = ",".join(metalog.lst)
 
     metalog = LineLogger()
@@ -330,13 +356,14 @@ def mccode_test(branchdir, testdir, limitinstrs=None, instrfilter=None, version=
     # displayname must be unique, we can return a dict, which eases comparison between tests
     obj = {}
     for t in tests:
-        obj[t.get_display_name()] = t.get_json_repr()
+        obj.update(t.json_dict())
     obj["_meta"] = metainfo
     return obj
 
 #
 # Utility
 #
+
 
 def activate_mccode_version(version, mccoderoot):
     '''
@@ -350,90 +377,83 @@ def activate_mccode_version(version, mccoderoot):
     os.environ["PATH"] = "%s/miniconda3/bin:%s/bin:%s" % (branchdir, branchdir, oldpath)
     return oldpath
 
+
 def deactivate_mccode_version(oldpath):
     ''' clean up path changes, restoring oldpath '''
     del os.environ["MCSTAS"]
     os.environ["PATH"] = oldpath
 
-def create_test_dir(testdir):
-    ''' just create testdir or exit '''
-    if not os.path.exists(testdir):
-        mkdir(testdir)
-    if not os.path.exists(testdir):
-        logging.info("could not create test folder, exiting...")
-        quit()
 
-def create_label_dir(testdir, label):
-    if not os.path.exists(testdir):
-        mkdir(testdir)
-    if not os.path.exists(testdir):
-        logging.info("could not create test folder, exiting...")
-        quit()
-    labeldir = join(testdir, label)
-    if not os.path.exists(labeldir):
-        mkdir(labeldir)
-    return labeldir
+def create_test_dir(test_dir: Path):
+    """ just create test_dir or exit """
+    if not test_dir.is_dir():
+        test_dir.mkdir(parents=True)
+
+
+def create_label_dir(test_dir: Path, label: str):
+    label_dir = test_dir.joinpath(label)
+    if not label_dir.is_dir():
+        label_dir.mkdir(parents=True)
+    return label_dir
+
 
 def create_datetime_testdir(testroot):
     datetime = utils.get_datetimestr()
-    create_label_dir(testroot, datetime)
-    return join(testroot, datetime)
+    return create_label_dir(testroot, datetime)
 
 #
 # Program functions for every main test mode
 #
 
-def run_default_test(testdir, mccoderoot, limit, instrfilter):
-    ''' tests the default mccode version '''
+def write_report(label_dir, version, results):
+    report_file = label_dir.joinpath(f"testresults_{version}.json")
+    with open(os.path.join(report_file), "w") as file:
+        file.write(json.dumps(results, indent=2))
+    logging.debug("")
+    logging.debug(f"Test results written to: {report_file}")
 
+
+def run_default_test(testdir: Path, mccoderoot: Path, limit, instrfilter):
+    """ tests the default mccode version """
     # get default/system version number
     logger = LineLogger()
-    utils.run_subtool_to_completion("%s --version" % (mccode_config.configuration["MCRUN"]), stdout_cb=logger.logline)
-    try:
-        version = logger.lst[-1].strip()
-    except:
-        logging.info("no 'mcstas --version' output, try using --configs")
+    cmd = f"{mccode_config.configuration['MCRUN']} --version"
+    utils.run_subtool_to_completion(cmd, stdout_cb=logger.logline)
+    if len(logger.lst) < 1:
+        logging.info(f"no '{cmd}' output, try using --configs")
         quit(1)
 
+    version = logger.lst[-1].strip()
     # create single-run test directory
-    labeldir = create_label_dir(testdir, version)
+    label_dir = create_label_dir(testdir, version)
 
-    logging.info("Testing: %s" % version)
+    logging.info(f"Testing: {version}")
     logging.info("")
-    results = mccode_test(os.path.join(mccoderoot, version), labeldir, limit, instrfilter)
-    
-    reportfile = os.path.join(labeldir, "testresults_%s.json" % version)
-    open(os.path.join(reportfile), "w").write(json.dumps(results, indent=2))
+    results = mccode_test(mccoderoot.joinpath(version), label_dir, limit, instrfilter)
 
-    logging.debug("")
-    logging.debug("Test results written to: %s" % reportfile)
+    write_report(label_dir, version, results)
 
 
-def run_version_test(testdir, mccoderoot, limit, instrfilter, version):
+def run_version_test(testdir: Path, mccoderoot: Path, limit, instrfilter, version):
     ''' as run_default_test, but activates/deactivates and ses a specific mccode version if it exists '''
 
     # verify that version exists
-    if not os.path.isfile(os.path.join(mccoderoot, version, "environment")):
-        print("mccode version %s could not be found, exiting..." % version)
+    if not mccoderoot.joinpath(version, "environment").is_file():
+        logging.info(f"mccode version {version} could not be found, exiting...")
         quit(1)
 
     # create single-run test directory
-    labeldir = create_label_dir(testdir, version)
+    label_dir = create_label_dir(testdir, version)
 
-    oldpath = activate_mccode_version(version, mccoderoot)
+    old_path = activate_mccode_version(version, mccoderoot)
     try:
-        logging.info("Testing: %s" % version)
+        logging.info(f"Testing: {version}")
         logging.info("")
-
-        results = mccode_test(os.path.join(mccoderoot, version), labeldir, limit, instrfilter, version)
+        results = mccode_test(mccoderoot.joinpath(version), label_dir, limit, instrfilter, version)
     finally:
-        deactivate_mccode_version(oldpath)
+        deactivate_mccode_version(old_path)
 
-    reportfile = os.path.join(labeldir, "testresults_%s.json" % version)
-    open(os.path.join(reportfile), "w").write(json.dumps(results, indent=2))
-
-    logging.debug("")
-    logging.debug("Test results written to: %s" % reportfile)
+    write_report(label_dir, version, results)
 
 
 def run_configs_test(testdir, mccoderoot, limit, configfilter, instrfilter):
@@ -481,63 +501,49 @@ def run_configs_test(testdir, mccoderoot, limit, configfilter, instrfilter):
 
     # test labels loop
     for f in get_config_files(configfilter):
-        [version,label] = extract_config_mccode_version(f)
+        [version, label] = extract_config_mccode_version(f)
 
         oldpath = activate_mccode_version(version, mccoderoot)
         try:
             #bckfile = activate_config(version, mccoderoot, f)
             try:
                 logging.info("")
-                label0=label
-                label=label+"_"+ncount
-                logging.info("Testing label: %s" % label)
-
-                # craete the proper test dir
-                labeldir = create_label_dir(testdir, label)
-                results = mccode_test(os.path.join(mccoderoot, version), labeldir, limit, instrfilter, label0)
-
+                label0 = label
+                label = label + "_" + ncount
+                logging.info(f"Testing label: {label}")
+                # create the proper test dir
+                label_dir = create_label_dir(testdir, label)
+                results = mccode_test(mccoderoot.joinpath(version), label_dir, limit, instrfilter, label0)
                 # write local test result
-                reportfile = os.path.join(labeldir, "testresults_%s.json" % label)
-                open(os.path.join(reportfile), "w").write(json.dumps(results, indent=2))
-            
-                logging.debug("")
-                logging.debug("Test results written to: %s" % reportfile)
+                write_report(label_dir, version, results)
             finally:
                 pass
         finally:
             deactivate_mccode_version(oldpath)
 
 
-def show_installed_versions(mccoderoot):
-    ''' utility function, prints identified mccode versions to console '''
+def show_installed_versions(mccoderoot: Path):
+    """ utility function, prints identified mccode versions to console """
 
     def print_to_console(str):
-        ''' used with popen wrapper '''
+        """ used with popen wrapper """
         logging.info(str)
 
-    logging.info("Test environment mode, using output of 'mcstas --vesion'")
+    flavour = mccode_config.configuration["MCCODE"]
+    logging.info(f"Test environment mode, using output of '{flavour} --vesion'")
     logging.info("")
 
-    dirnames = []
-    branchnames = []
-    for (_, dirnames, _) in os.walk(mccoderoot):
-        break
-    for d in dirnames:
-        for (_, _, files) in os.walk(join(mccoderoot, d)):
-            break
-        if "environment" in files:
-            branchnames.append(d)
+    # collect all folders which contain an 'environment' file
+    branchnames = [x.parent for x in mccoderoot.rglob('environment') if x.is_file()]
 
-    for v in branchnames:
+    for branchdir in branchnames:
         # test environment
-        branchdir = join(mccoderoot, v)
-    
-        os.environ["MCSTAS"] = branchdir
-        oldpath = os.environ["PATH"]
-        os.environ["PATH"] = "%s/miniconda3/bin:%s/bin:%s" % (branchdir, branchdir, oldpath)
+        os.environ["MCSTAS"] = str(branchdir)
+        old_path = os.environ["PATH"]
+        os.environ['PATH'] = f"{branchdir.joinpath('miniconda3','bin')}:{branchdir.joinpath('bin')}:{old_path}"
     
         # run the mcstas --version command
-        cmd = "mcstas --version"
+        cmd = f"{flavour} --version"
         utils.run_subtool_to_completion(cmd, stdout_cb=print_to_console, stderr_cb=print_to_console)
         logging.info("")
     
@@ -545,14 +551,16 @@ def show_installed_versions(mccoderoot):
     
         # restore environment
         del os.environ["MCSTAS"]
-        os.environ["PATH"] = oldpath
+        os.environ["PATH"] = old_path
 
-    logging.info("Selectable version names are: %s" % ", ".join(branchnames))
+    version_names = ','.join([x.name for x in branchnames])
+    logging.info("Selectable version names are: {version_names}")
     logging.info("")
 
 ncount = None
 mpi = None
 openacc = None
+
 
 def main(args):
     # mutually excusive main branches
@@ -578,24 +586,29 @@ def main(args):
     else:
         logging.basicConfig(level=logging.INFO, format="%(message)s")
     if not testroot:
-        testroot = "/tmp/mctest"
+        testroot = Path("/tmp/mctest")
     if not testdir:
         testdir = create_datetime_testdir(testroot)
-        logging.debug("Using test root:         %s" % testroot)
+        logging.debug(f"Using test root:         {testroot}")
     else:
+        if not isinstance(testdir, Path) and isinstance(testdir, str):
+            testdir = Path(testdir)
         create_test_dir(testdir)
-        logging.debug("Using explicit test dir: %s" % testdir)
+        logging.debug(f"Using explicit test dir: {testdir}")
 
     if not mccoderoot:
-        mccoderoot = "/usr/share/mcstas/"
-    if not os.path.exists(mccoderoot):
+        mccoderoot = Path("/usr/share/mcstas/")
+    if isinstance(mccoderoot, str):
+        mccoderoot = Path(mccoderoot)
+    if not mccoderoot.is_dir():
         logging.info("mccoderoot does not exist")
         quit(1)
-    logging.debug("Using mccode root:       %s" % mccoderoot)
+    logging.debug(f"Using mccode root:       {mccoderoot}")
     if limit:
         try:
             limit = int(args.limit[0])
-        except:
+        except ValueError as ex:
+            logging.debug(f"{ex}")
             logging.info("--limit must be a number")
             quit(1)
     logging.debug("")
@@ -605,10 +618,10 @@ def main(args):
         ncount = args.ncount[0]
     else:
         ncount = "1e6"
-    logging.info("ncount is: %s" % ncount)
+    logging.info(f"ncount is: {ncount}")
     if args.mpi:
         mpi = args.mpi[0]
-        logging.info("mpi count is: %s" % mpi)
+        logging.info(f"mpi count is: {mpi}")
     if args.openacc:
         openacc = True
         logging.info("openacc is enabled")
@@ -629,20 +642,23 @@ def main(args):
 
 
 if __name__ == '__main__':
+    flavour = mccode_config.configuration['MCCODE']
     parser = argparse.ArgumentParser(description=__doc__)
     parser.add_argument('testversion', nargs="?", help='mccode version to test')
-    parser.add_argument('--ncount', nargs=1, help='ncount sent to %s' % (mccode_config.configuration["MCRUN"]) )
-    parser.add_argument('--mpi', nargs=1, help='mpi nodecount sent to %s' % (mccode_config.configuration["MCRUN"]) )
-    parser.add_argument('--openacc', action='store_true', help='openacc flag sent to %s' % (mccode_config.configuration["MCRUN"]))
-    parser.add_argument('--configs', action='store_true', help='test config files under mccodelib/MCCODE')
-    parser.add_argument('--config', nargs="?", help='test this specific config only - label name or absolute path (enables --configs)')
-    parser.add_argument('--instr', nargs="?", help='test only intruments matching this filter (py regex)')
+    parser.add_argument('--ncount', nargs=1, help=f'ncount sent to {flavour}')
+    parser.add_argument('--mpi', nargs=1, help=f'mpi nodecount sent to {flavour}')
+    parser.add_argument('--openacc', action='store_true', help=f'openacc flag sent to {flavour}')
+    parser.add_argument('--configs', action='store_true', help=f'test config files under mccodelib/{flavour}-test')
+    parser.add_argument('--config', nargs="?",
+                        help='test this specific config only - label name or absolute path (enables --configs)')
+    parser.add_argument('--instr', nargs="?", help='test only instruments matching this filter (py regex)')
     parser.add_argument('--mccoderoot', nargs='?', help='manually select root search folder for mccode installations')
     parser.add_argument('--testroot', nargs='?', help='output test results in a datetime folder in this root')
     parser.add_argument('--testdir', nargs='?', help='output test results directly in this dir (overrides testroot)')
-    parser.add_argument('--limit', nargs=1, help='test only the first [LIMIT] instrs in every version')
+    parser.add_argument('--limit', nargs=1, help='test only the first [LIMIT] instruments in every version')
     parser.add_argument('--versions', action='store_true', help='display local versions info')
-    parser.add_argument('--verbose', action='store_true', help='output a test/notest instrument status header before each test')
+    parser.add_argument('--verbose', action='store_true',
+                        help='output a test/notest instrument status header before each test')
     parser.add_argument('--skipnontest', action='store_true', help='Skip compilation of instruments without a test')
 
     args = parser.parse_args()
